% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_and_label_order_476.R
\name{catalog_and_label_order_476}
\alias{catalog_and_label_order_476}
\title{Return the 476-channel indel type classification table}
\usage{
catalog_and_label_order_476()
}
\value{
A 476-row data frame with columns: IndelType, Indel.
}
\description{
Constructs a data frame mapping 476 indel channels to their type and class.
}
\keyword{internal}
