% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unstaple_SBS96_rownames.R
\name{unstaple_SBS96_rownames}
\alias{unstaple_SBS96_rownames}
\title{Convert SBS96-channel mutations-type identifiers like this \code{"A[C>A]T" -> "ACTA"}}
\usage{
unstaple_SBS96_rownames(c1)
}
\arguments{
\item{c1}{A vector of character strings with the mutation indicated by
e.g. \code{[C>A]} in the middle.}
}
\description{
Convert SBS96-channel mutations-type identifiers like this \code{"A[C>A]T" -> "ACTA"}
}
\keyword{internal}
