% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_83.R
\name{plot_83}
\alias{plot_83}
\title{Plot 83-type indel catalog using ggplot2 and return the ggplot object}
\usage{
plot_83(
  catalog,
  text_size = 3,
  plot_title = colnames(catalog)[1],
  grid = TRUE,
  upper = TRUE,
  xlabels = TRUE,
  ylabels = TRUE,
  ylim = NULL,
  base_size = 11
)
}
\arguments{
\item{catalog}{An 83-row, 1 - column indel catalog}

\item{text_size}{Numeric. Size of text labels in the plot. Default is 3.}

\item{plot_title}{Character. Title displayed above the plot. Default is "test".}

\item{grid}{Logical, draw grid lines}

\item{upper}{Logical, draw category labels above bars}

\item{xlabels}{Logical, draw x-axis labels}

\item{ylabels}{Logical, draw y-axis labels}

\item{ylim}{Optional y-axis limits}

\item{base_size}{Base font size for ggplot2's theme.}
}
\value{
A ggplot object
}
\description{
Plot 83-type indel catalog using ggplot2 and return the ggplot object
}
